+++ spring-boot/src/main/java/com/morimoto/taichi/moruku/domain/entity/Practice.java
@@ -23,7 +23,7 @@ public class Practice {
     private String title;
     private String description;
     private Integer maxParticipants;
-    private byte prefectureId;
+    private Integer prefectureId;
     private UUID organizerId;
     private LocalDate heldOn;
     private LocalDateTime createdAt;
